refactor recipe.js
