+++ Code/backend/dao/recipes/index.js
@@ -1,3 +1,4 @@
 module.exports.AddNewRecipe = require('./addNewRecipe');
 module.exports.GetAllRecipes = require('./getAllRecipe');
 module.exports.GetTotalRecipeCount = require('./getTotalRecipeCount');
+module.exports.UpdateRecipe = require('./updateRecipe');
